♻️ move service_factory out of grpc package
