+++ caikit/runtime/__init__.py
@@ -13,8 +13,4 @@
 # limitations under the License.
 
 # Local
-from .grpc_server.service_factory import (
-    get_inference_request,
-    get_train_params,
-    get_train_request,
-)
+from .service_factory import get_inference_request, get_train_params, get_train_request
