+++ caikit/runtime/grpc_server/servicers/global_train_servicer.py
@@ -29,8 +29,8 @@ from caikit import get_config
 from caikit.core import MODEL_MANAGER, ModuleBase
 from caikit.interfaces.common.data_model.stream_sources import S3Path
 from caikit.interfaces.runtime.data_model import TrainingJob
-from caikit.runtime.grpc_server.service_factory import ServicePackage
 from caikit.runtime.model_management.model_manager import ModelManager
+from caikit.runtime.service_factory import ServicePackage
 from caikit.runtime.service_generation.rpcs import ModuleClassTrainRPC
 from caikit.runtime.types.caikit_runtime_exception import CaikitRuntimeException
 from caikit.runtime.utils.import_util import clean_lib_names, get_data_model
