+++ tests/runtime/grpc_server/conftest.py
@@ -22,10 +22,6 @@ import alog
 from caikit.core import MODEL_MANAGER
 from caikit.core.data_model.dataobject import render_dataobject_protos
 from caikit.runtime.grpc_server.grpc_server import RuntimeGRPCServer
-from caikit.runtime.grpc_server.service_factory import (
-    ServicePackage,
-    ServicePackageFactory,
-)
 from caikit.runtime.grpc_server.servicers.global_predict_servicer import (
     GlobalPredictServicer,
 )
@@ -33,6 +29,7 @@ from caikit.runtime.grpc_server.servicers.global_train_servicer import (
     GlobalTrainServicer,
 )
 from caikit.runtime.model_management.loaded_model import LoadedModel
+from caikit.runtime.service_factory import ServicePackage, ServicePackageFactory
 from caikit.runtime.service_generation.rpcs import TaskPredictRPC
 from tests.conftest import temp_config
